+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [[#250]](https://github.com/Marcin-Migdal/m-component-library/issues/250) **[PATCH]** Improve DropdownMenu to use DropdownSubMenu with template
 - [[#245]](https://github.com/Marcin-Migdal/m-component-library/issues/245) **[PATCH]** Delete number form Accordion sectionId type, simplify logic
+- [[#232]](https://github.com/Marcin-Migdal/m-component-library/issues/232) **[PATCH]** Add Responsive behavior to the Breadcrumb when it does not fit in its container
 
 ### Release date 04.07.2025 - Release 2.1.5